refactor: change timing fields from float to double
.

+++ ParallelCollisionLab/Collision/CollisionTypes.h
@@ -12,26 +12,20 @@ struct FCollisionPair
     int IndexB = -1;
 };
 
-//=============================================================================
-// FCollisionManifold - Narrow Phase output: contact geometry data
-//=============================================================================
 struct FCollisionManifold
 {
     int      IndexA      = -1;
     int      IndexB      = -1;
-    FVector3 Normal      = FVector3(0.0f, 0.0f, 0.0f); // Unit normal pointing from B to A
-    float    Penetration = 0.0f;                       // Overlap depth (> 0 when colliding)
+    FVector3 Normal      = FVector3(0.0f, 0.0f, 0.0f);
+    float    Penetration = 0.0f;
 };
 
-//=============================================================================
-// FCollisionStats - Profiling and pipeline metrics per simulation tick
-//=============================================================================
 struct FCollisionStats
 {
-    float    BroadPhaseTimeMs     = 0.0f;
-    float    NarrowPhaseTimeMs    = 0.0f;
-    float    ResolutionTimeMs     = 0.0f;
-    float    TotalSolveTimeMs     = 0.0f;
+    double   BroadPhaseTimeMs     = 0.0;
+    double   NarrowPhaseTimeMs    = 0.0;
+    double   ResolutionTimeMs     = 0.0;
+    double   TotalSolveTimeMs     = 0.0;
     uint64_t CandidatePairCount   = 0;
     uint64_t ActualCollisionCount = 0;
 };
